Final Until Further Improvements

+++ app/Http/Controllers/Backend/BrandController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\BrandDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Brand;
+use App\Models\Product;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
 use Str;
@@ -106,10 +107,16 @@ class BrandController extends Controller
     /**
      * Remove the specified resource from storage.
      */
+
     public function destroy(string $id)
     {
         $brand = Brand::findOrFail($id);
+        if(Product::where('brand_id', $brand->id)->count() > 0){
+            return response()->json(['status' => 'error', 'message' => "This Brand contains Products. You can't delete it."]);
+        }
         $brand->delete();
+
+        return response()->json(['message' => 'Brand has been deleted successfully.']);
     }
 
     public function changeStatus(Request $request)
